Move helper function to helper file
